pyproject.toml
[project]
name = "100-knocks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ginza>=5.2.0",
    "ja-ginza>=5.2.0",
    "mecab-python3>=1.0.10",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "spacy>=3.8.11",
]
